Add queries to fetch data
Added queries to fetch data which get displayed on the page on first
load. All data from database are fetched and rendered.

For the search and filter fields, a separate call is made to fetch the
necessary data. It would have been optimal to use the same data displayed
on the page, however, when data is searched and filtered, the data shown
in the dropdowns are determined by the search and filter results, which
may not be desirable.

Jobs can be searched by jot title, while filtering can be done by company,
city and investors.